Conflicts:
	src/cc/AsyncComm/ApplicationHandler.h
	src/cc/AsyncComm/Comm.cc
	src/cc/AsyncComm/Comm.h
	src/cc/AsyncComm/CommAddress.h
	src/cc/AsyncComm/ConnectionManager.h
	src/cc/AsyncComm/Event.h
	src/cc/AsyncComm/HandlerMap.h
	src/cc/AsyncComm/IOHandler.h
	src/cc/AsyncComm/IOHandlerAccept.h
	src/cc/AsyncComm/IOHandlerData.cc
	src/cc/AsyncComm/IOHandlerData.h
	src/cc/AsyncComm/IOHandlerDatagram.cc
	src/cc/AsyncComm/IOHandlerDatagram.h
	src/cc/AsyncComm/PollTimeout.h
	src/cc/AsyncComm/Reactor.cc
	src/cc/AsyncComm/Reactor.h
	src/cc/AsyncComm/ReactorFactory.h
	src/cc/AsyncComm/ReactorRunner.cc
	src/cc/Common/Allocator.h
	src/cc/Common/Logger.cc
	src/cc/Common/Logger.h
	src/cc/Common/atomic.h
	src/cc/Common/tests/failure_inducer_test.cc
	src/cc/DfsBroker/local/main.cc
	src/cc/Hypertable/Lib/IntervalScanner.cc
	src/cc/Hypertable/Lib/TableMutatorAsync.cc
	src/cc/Hypertable/Lib/TableMutatorAsync.h
	src/cc/Hypertable/Lib/Types.h
	src/cc/Hypertable/Lib/tests/name_id_mapper_test.cc
	src/cc/Hypertable/Master/RangeServerConnection.cc
	src/cc/Hypertable/RangeServer/CellStoreBlockIndexMap.h
	src/cc/Hypertable/RangeServer/MaintenancePrioritizer.cc
	src/cc/Hypertable/RangeServer/MaintenanceScheduler.cc
	src/cc/Hypertable/RangeServer/RangeServer.cc
	src/cc/Hypertable/RangeServer/RequestHandlerShutdown.cc
	src/cc/Hypertable/RangeServer/TimerHandler.cc
	src/cc/Hypertable/RangeServer/TimerHandler.h
	src/cc/Hypertable/RangeServer/main.cc
	src/cc/Tools/Lib/CommandShell.cc
	src/cc/Tools/hyperspace/test/hyperspaceTest.cc
	src/cc/Tools/hyperspace/test/hyperspaceTest.cfg

+++ src/cc/Tools/load_generator/QueryThread.cc
@@ -76,14 +76,14 @@ void QueryThread::operator()() {
       }
       delete scanner;
 
-      double latency = 0.0;
+      int64_t latency = 0;
       stop_clocks = clock();
       if (stop_clocks < start_clocks)
         latency += (int64_t)((double)((std::numeric_limits<clock_t>::max() - start_clocks) + stop_clocks) / clocks_per_usec);
       else
         latency += (int64_t)((double)(stop_clocks - start_clocks) / clocks_per_usec);
       m_state.cum_latency += latency;
-      m_state.cum_sq_latency += ::pow(latency, 2);
+      m_state.cum_sq_latency += ::pow(latency, 2.);
       if (latency < m_state.min_latency)
         m_state.min_latency = latency;
       if (latency > m_state.max_latency)
